Fix: Home BottomNavigation Icon Error
