+++ app/src/main/res/drawable/bottom_navi_home.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:drawable="@drawable/ic_home2"
+        android:drawable="@drawable/ic_home_color"
         android:state_checked="true" />
 
     <item
